move files to moc (linux/mac)
