+++ modules/text_editor/src/cpp/QtTextEdit.h
